created a new component to understand the code in a simple way
